New import strategy for foreign vocabs
Till now the URI of a term from a foreign vocabulary would be
introduced in the local triple store according to the path definitions.
This has as major drawback that you cannot use the standard WissKI/Drupal view
for the term as it has a non-local URI prefix.
The new behaviour introduces a local URI for each foreign term that is used
locally. This local instance may be treated like a normal local instance.
Furthermore it is linked to the foreign term URI with a selected property,
so that the relation to the term is clear.

Update notice: The db table wisski_vocab_ctrl_vocabularies has changed!